move stuff to conftest
